migrate over to the new loop manager
