+++ flytekit/utils/asyn.py
@@ -13,7 +13,9 @@ import asyncio
 import atexit
 import os
 import threading
+from concurrent.futures import ThreadPoolExecutor
 from contextlib import contextmanager
+from contextvars import copy_context
 from typing import Any, Awaitable, Callable, TypeVar
 
 from flytekit.loggers import logger
@@ -99,3 +101,13 @@ class _AsyncLoopManager:
 
 loop_manager = _AsyncLoopManager()
 run_sync = loop_manager.run_sync
+
+
+class ContextExecutor(ThreadPoolExecutor):
+    def __init__(self):
+        self.context = copy_context()
+        super().__init__(initializer=self._set_child_context)
+
+    def _set_child_context(self):
+        for var, value in self.context.items():
+            var.set(value)
